requirements.txt
feedparser
requests
sgmllib3k
python-dotenv
